+++ app/Repositories/Services/ServicesRepositoryEloquent.php
@@ -10,10 +10,12 @@ use Illuminate\Http\Request;
 class ServicesRepositoryEloquent implements ServicesRepositoryInterface
 {
     private $service;
+    private $serviceValidate;
 
     public function __construct(Service $service)
     {
         $this->service = $service;
+        $this->serviceValidate = auth()->user()->service;
     }
 
     public function getAll()
@@ -33,40 +35,58 @@ class ServicesRepositoryEloquent implements ServicesRepositoryInterface
 
     public function store(Request $request)
     {
-        return $this->service->create($request->all());
+        if ($this->serviceValidate == -1) {
+            return $this->service->create($request->all());
+        }
+        return [];
     }
 
     public function storeCategory(Request $request)
     {
-        return Category::create($request->all());
+        if ($this->serviceValidate == -1) {
+            return Category::create($request->all());
+        }
+        return [];
     }
 
     public function update(int $id, Request $request)
     {
-        return $this->service
-            ->where('id', $id)
-            ->update($request->all());
+        if ($this->serviceValidate == -1) {
+            return $this->service
+                ->where('id', $id)
+                ->update($request->all());
+        }
+        return [];
     }
 
     public function updateCategory(int $id, int $category_id, Request $request)
     {
-        return Category::where('id', $category_id)
-            ->where('service_id', $id)
-            ->update($request->all());
+        if ($this->serviceValidate == -1) {
+            return Category::where('id', $category_id)
+                ->where('service_id', $id)
+                ->update($request->all());
+        }
+        return [];
     }
 
     public function destroy(int $id)
     {
-        $service = $this->service->find($id);
-        $service->category()->delete();
-        return $service->delete();
+        if ($this->serviceValidate == -1) {
+            $service = $this->service->find($id);
+            $service->category()->delete();
+            return $service->delete();
+        }
+        return [];
     }
 
     public function destroyCategory(int $id, int $category_id)
     {
-        $category = Category::where('id', $category_id)
-            ->where('service_id', $id)
-            ->first();
-        return $category->delete();
+        if ($this->serviceValidate == -1) {
+            $category = Category::where('id', $category_id)
+                ->where('service_id', $id)
+                ->first();
+            return $category->delete();
+        }
+        return [];
     }
 }
